ci: Split e2e tests to make them faster (#4404)

* ok
