+++ test/e2e/argo_server_test.go
@@ -1,10 +1,9 @@
-// +build e2e
+// +build cli
 
 package e2e
 
 import (
 	"bufio"
-	"crypto/tls"
 	"io/ioutil"
 	"net/http"
 	"strings"
@@ -28,13 +27,6 @@ import (
 
 const baseUrl = "http://localhost:2746"
 
-var httpClient = &http.Client{
-	Transport: &http.Transport{TLSClientConfig: &tls.Config{InsecureSkipVerify: true}},
-	CheckRedirect: func(req *http.Request, via []*http.Request) error {
-		return http.ErrUseLastResponse
-	},
-}
-
 // ensure basic HTTP functionality works,
 // testing behaviour really is a non-goal
 type ArgoServerSuite struct {
@@ -47,13 +39,6 @@ func (s *ArgoServerSuite) BeforeTest(suiteName, testName string) {
 	var err error
 	s.bearerToken, err = s.GetServiceAccountToken()
 	s.CheckError(err)
-}
-
-type httpLogger struct {
-}
-
-func (d *httpLogger) Logf(fmt string, args ...interface{}) {
-	log.Debugf(fmt, args...)
 }
 
 func (s *ArgoServerSuite) e() *httpexpect.Expect {
