+++ test/e2e/cli_with_server_test.go
@@ -1,4 +1,4 @@
-// +build e2e
+// +build cli
 
 package e2e
