+++ test/e2e/client.go
@@ -1,0 +1,13 @@
+package e2e
+
+import (
+	"crypto/tls"
+	"net/http"
+)
+
+var httpClient = &http.Client{
+	Transport: &http.Transport{TLSClientConfig: &tls.Config{InsecureSkipVerify: true}},
+	CheckRedirect: func(req *http.Request, via []*http.Request) error {
+		return http.ErrUseLastResponse
+	},
+}
