+++ test/e2e/http_logger.go
@@ -1,0 +1,10 @@
+package e2e
+
+import log "github.com/sirupsen/logrus"
+
+type httpLogger struct {
+}
+
+func (d *httpLogger) Logf(fmt string, args ...interface{}) {
+	log.Debugf(fmt, args...)
+}
